requirements.txt
nonebot-adapter-onebot==2.4.6
nonebot-plugin-apscheduler==0.5.0
nonebot-plugin-htmlrender==0.6.5
nonebot-plugin-localstore==0.7.1
nonebot-plugin-orm==0.7.6
nonebot2==2.4.2
nonechat==0.3.0
noneprompt==0.1.9
nonestorage==0.1.0
typing-inspection==0.4.0
typing_extensions==4.12.2
aiosqlite==0.20.0
APScheduler==3.10.4
DateTime==5.5
feedparser==6.0.11
fastapi==0.115.0
httpx==0.27.2
idna==3.10
ImageHash==4.3.2
inquirerpy==0.3.4
installer==0.7.0
irc3==1.1.10
jaraco.context==6.0.1
jaraco.functools==4.1.0
jsonpointer==2.1
keyring==25.6.0
lfmh==1.1.1
loguru==0.7.2
lxml==5.3.0
magneturi==1.3
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
msgpack==1.1.0
multidict==6.1.0
musicbrainzngs==0.7.1
nb-cli==1.4.2
nb-cli-plugin-docker==0.4.1
nonebot-plugin-apscheduler==0.5.0
nonebot-plugin-htmlrender==0.6.5
nonebot-plugin-localstore==0.7.1
nonebot-plugin-orm==0.7.6
nonebot2==2.4.2
nonechat==0.3.0
noneprompt==0.1.9
nonestorage==0.1.0
onebot==1.3.3
SQLAlchemy==2.0.35
pydantic==2.9.2
typing-inspection==0.4.0
typing_extensions==4.12.2
requests==2.32.3